AI_Farmtopia/game_engine.py
```python
import json
from ai_components import AIComponentSchema
from farming_companions import FarmingCompanion
from ecosystem import Ecosystem
from personalized_gameplay import PersonalizedGameplay
from weather_system import WeatherSystem
from quests import Quest
from farming_whiz import FarmingWhiz
from monetization import Monetization

class GameEngine:
    def __init__(self):
        self.playerData = {}
        self.gameSettings = {}
        self.aiComponents = AIComponentSchema()
        self.weatherData = {}
        self.questData = {}
        self.monetizationData = {}

    def loadGame(self):
        with open('playerData.json', 'r') as file:
            self.playerData = json.load(file)
        with open('gameSettings.json', 'r') as file:
            self.gameSettings = json.load(file)
        self.aiComponents.loadComponents()
        self.weatherData = WeatherSystem().getWeatherData()
        self.questData = Quest().getQuestData()
        self.monetizationData = Monetization().getMonetizationData()

    def saveGame(self):
        with open('playerData.json', 'w') as file:
            json.dump(self.playerData, file)
        with open('gameSettings.json', 'w') as file:
            json.dump(self.gameSettings, file)
        self.aiComponents.saveComponents()
        WeatherSystem().saveWeatherData(self.weatherData)
        Quest().saveQuestData(self.questData)
        Monetization().saveMonetizationData(self.monetizationData)

    def updatePlayer(self, newPlayerData):
        self.playerData.update(newPlayerData)
        self.saveGame()

    def updateSettings(self, newGameSettings):
        self.gameSettings.update(newGameSettings)
        self.saveGame()

    def updateAI(self, newAIComponents):
        self.aiComponents.update(newAIComponents)
        self.saveGame()

    def updateWeather(self, newWeatherData):
        self.weatherData.update(newWeatherData)
        self.saveGame()

    def updateQuests(self, newQuestData):
        self.questData.update(newQuestData)
        self.saveGame()

    def updateMonetization(self, newMonetizationData):
        self.monetizationData.update(newMonetizationData)
        self.saveGame()

    def displayAd(self):
        ad = Monetization().getAd()
        return ad

    def purchasePowerUp(self, powerUpId):
        powerUp = Monetization().purchasePowerUp(powerUpId)
        return powerUp
```