AI_Farmtopia/tests/test_game_engine.py
```python
import unittest
from AI_Farmtopia.game_engine import loadGame, saveGame, updatePlayer, updateSettings, updateAI, updateWeather, updateQuests, updateMonetization, displayAd, purchasePowerUp

class TestGameEngine(unittest.TestCase):

    def setUp(self):
        self.playerData = {
            'name': 'Test Player',
            'level': 1,
            'progress': 0,
            'preferences': {}
        }
        self.gameSettings = {
            'difficulty': 'easy',
            'sound': True,
            'notifications': True
        }
        self.aiComponents = []
        self.weatherData = {}
        self.questData = []
        self.monetizationData = {
            'purchases': [],
            'ads': []
        }

    def test_loadGame(self):
        result = loadGame()
        self.assertIsNotNone(result)

    def test_saveGame(self):
        result = saveGame(self.playerData, self.gameSettings, self.aiComponents, self.weatherData, self.questData, self.monetizationData)
        self.assertTrue(result)

    def test_updatePlayer(self):
        newPlayerData = updatePlayer(self.playerData)
        self.assertIsNotNone(newPlayerData)

    def test_updateSettings(self):
        newGameSettings = updateSettings(self.gameSettings)
        self.assertIsNotNone(newGameSettings)

    def test_updateAI(self):
        newAIComponents = updateAI(self.aiComponents)
        self.assertIsNotNone(newAIComponents)

    def test_updateWeather(self):
        newWeatherData = updateWeather(self.weatherData)
        self.assertIsNotNone(newWeatherData)

    def test_updateQuests(self):
        newQuestData = updateQuests(self.questData)
        self.assertIsNotNone(newQuestData)

    def test_updateMonetization(self):
        newMonetizationData = updateMonetization(self.monetizationData)
        self.assertIsNotNone(newMonetizationData)

    def test_displayAd(self):
        result = displayAd()
        self.assertTrue(result)

    def test_purchasePowerUp(self):
        result = purchasePowerUp('speed_boost')
        self.assertTrue(result)

if __name__ == '__main__':
    unittest.main()
```